Moved some of the fetching, rendering, and components around to advocate for fetching only within the page
