+++ apps/www/src/app/articles/page.tsx
@@ -1,7 +1,10 @@
+import { sanityClient } from '@lib/sanity.client'
+import { articlesPageQuery } from '@lib/queries'
+
 import { SectionContainer } from '@components/section-container'
 import { SectionTitle } from '@components/section-title'
 import { Text } from '@components/text'
-import { ArticlesPostGrid } from './articles-post-grid'
+import { ArticlesPostGrid } from '../components/pages/articles'
 
 import type { Metadata } from 'next'
 
@@ -10,7 +13,27 @@ export const metadata: Metadata = {
   description: 'Recent articles and learning materials from Downbeat Academy.'
 }
 
+async function getArticles() {
+  const res = sanityClient.fetch(
+    articlesPageQuery,
+    { 
+      next: {
+        revalidate: 60,
+      },
+    }
+  )
+
+  if (!res) {
+    throw new Error('Failed to fetch data.');
+  }
+
+  return res;
+}
+
 export default async function ArticlesPage() {
+
+  const articles = await getArticles();
+
   return (
     <>
       <SectionContainer>
@@ -25,7 +48,7 @@ export default async function ArticlesPage() {
           }
         />
         {/* @ts-expect-error Server Component */}
-        <ArticlesPostGrid />
+        <ArticlesPostGrid articles={articles} />
       </SectionContainer>
     </>
   )
