+++ apps/www/src/app/components/pages/articles/index.ts
@@ -1,0 +1,1 @@
+export { ArticlesPostGrid } from './articles-post-grid'
