+++ apps/www/src/app/components/pages/contributors/index.ts
@@ -1,0 +1,1 @@
+export { ContributorsGrid } from './contributors-grid'
